openwrt: Package both cyassl and openssl
This is alternative openwrt package makefile.  It drops the example
applications, as they're 600k+ and it's tedious to get them to build for
both variants of the ssl library.  It includes packages for both cyassl
and openssl variants, though the cyassl variant is only compile tested.
The notes in the original file about needing to add
--enable-opensslextra don't seem to get in the way of building at least?

If this all works out, I can get this into standard openwrt package
repositories.