Fix the travis.yml to build everything
When fixing the indentation for the coverity chagnes in 924bcfbc995e5c862216f6ad47a734a2c8f14370
the build matrix got turned off. This re-enables this.

Also change from using gcc/clang. What we want is to build on both Linux
and OSX. Not just run Linux with gcc+clang.

However to enable multi-os support one has to mail the travis support:
http://docs.travis-ci.com/user/multi-os/

"This feature needs to be enabled manually. If you wish to take advantage
of this feature, please send email to support@travis-ci.com, indicating
which repository you would like to have it enabled."